backend/python-ml-api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tensorflow>=2.6.0
pillow>=8.3.1
python-dotenv>=0.19.0
numpy>=1.19.5